API/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
motor
pydantic
pandas
numpy
spacy
unidecode
networkx
scikit-learn